requirements.txt
coverage==6.2
Django==2.2.10
django-debug-toolbar==3.2.4
django-debug-toolbar-force==0.1.8
django-nine==0.2.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
